requirements.txt
antlr4-python3-runtime==4.13.1
antlr4-tools==0.2.1
install-jdk==1.1.0
